requirements.txt
boto3==1.26.50
botocore==1.29.50
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
image-helpers==0.1.4
jmespath==1.0.1
Pillow==9.4.0
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
